+++ cmd/protoc-gen-doc/main_test.go
@@ -1,23 +1,14 @@
 package main_test
 
 import (
-	"github.com/stretchr/testify/suite"
-
 	"bytes"
 	"testing"
 
-	"github.com/pseudomuto/protoc-gen-doc/cmd/protoc-gen-doc"
+	. "github.com/pseudomuto/protoc-gen-doc/cmd/protoc-gen-doc"
+	"github.com/stretchr/testify/require"
 )
 
-type MainTest struct {
-	suite.Suite
-}
-
-func TestMain(t *testing.T) {
-	suite.Run(t, new(MainTest))
-}
-
-func (assert *MainTest) TestHandleFlags() {
+func TestHandleFlags(t *testing.T) {
 	tests := []struct {
 		args   []string
 		result bool
@@ -28,7 +19,7 @@ func (assert *MainTest) TestHandleFlags() {
 	}
 
 	for _, test := range tests {
-		f := main.ParseFlags(new(bytes.Buffer), test.args)
-		assert.Equal(test.result, main.HandleFlags(f))
+		f := ParseFlags(new(bytes.Buffer), test.args)
+		require.Equal(t, test.result, HandleFlags(f))
 	}
 }