+++ filters_test.go
@@ -1,21 +1,14 @@
 package gendoc_test
 
 import (
-	"github.com/pseudomuto/protoc-gen-doc"
-	"github.com/stretchr/testify/suite"
 	html "html/template"
 	"testing"
+
+	. "github.com/pseudomuto/protoc-gen-doc"
+	"github.com/stretchr/testify/require"
 )
 
-type FilterTest struct {
-	suite.Suite
-}
-
-func TestFilter(t *testing.T) {
-	suite.Run(t, new(FilterTest))
-}
-
-func (assert *FilterTest) TestPFilter() {
+func TestPFilter(t *testing.T) {
 	tests := map[string]string{
 		"Some content.":                          "<p>Some content.</p>",
 		"Some content.\nRight here.":             "<p>Some content.</p><p>Right here.</p>",
@@ -25,11 +18,11 @@ func (assert *FilterTest) TestPFilter() {
 	}
 
 	for input, output := range tests {
-		assert.Equal(html.HTML(output), gendoc.PFilter(input))
+		require.Equal(t, html.HTML(output), PFilter(input))
 	}
 }
 
-func (assert *FilterTest) TestParaFilter() {
+func TestParaFilter(t *testing.T) {
 	tests := map[string]string{
 		"Some content.":                          "<para>Some content.</para>",
 		"Some content.\nRight here.":             "<para>Some content.</para><para>Right here.</para>",
@@ -39,11 +32,11 @@ func (assert *FilterTest) TestParaFilter() {
 	}
 
 	for input, output := range tests {
-		assert.Equal(output, gendoc.ParaFilter(input))
+		require.Equal(t, output, ParaFilter(input))
 	}
 }
 
-func (assert *FilterTest) TestNoBrFilter() {
+func TestNoBrFilter(t *testing.T) {
 	tests := map[string]string{
 		"My content":                     "My content",
 		"My content \r\nHere.":           "My content Here.",
@@ -53,6 +46,6 @@ func (assert *FilterTest) TestNoBrFilter() {
 	}
 
 	for input, output := range tests {
-		assert.Equal(output, gendoc.NoBrFilter(input))
+		require.Equal(t, output, NoBrFilter(input))
 	}
 }