+++ plugin_test.go
@@ -1,25 +1,16 @@
 package gendoc_test
 
 import (
-	"github.com/golang/protobuf/proto"
-	"github.com/golang/protobuf/protoc-gen-go/plugin"
-	"github.com/stretchr/testify/suite"
-
 	"regexp"
 	"testing"
 
-	"github.com/pseudomuto/protoc-gen-doc"
+	"github.com/golang/protobuf/proto"
+	plugin_go "github.com/golang/protobuf/protoc-gen-go/plugin"
+	. "github.com/pseudomuto/protoc-gen-doc"
+	"github.com/stretchr/testify/require"
 )
 
-type PluginTest struct {
-	suite.Suite
-}
-
-func TestPlugin(t *testing.T) {
-	suite.Run(t, new(PluginTest))
-}
-
-func (assert *PluginTest) TestParseOptionsForBuiltinTemplates() {
+func TestParseOptionsForBuiltinTemplates(t *testing.T) {
 	results := map[string]string{
 		"docbook":  "output.xml",
 		"html":     "output.html",
@@ -31,45 +22,45 @@ func (assert *PluginTest) TestParseOptionsForBuiltinTemplates() {
 		req := new(plugin_go.CodeGeneratorRequest)
 		req.Parameter = proto.String(kind + "," + file)
 
-		options, err := gendoc.ParseOptions(req)
-		assert.Nil(err)
+		options, err := ParseOptions(req)
+		require.NoError(t, err)
 
-		renderType, err := gendoc.NewRenderType(kind)
-		assert.Nil(err)
+		renderType, err := NewRenderType(kind)
+		require.NoError(t, err)
 
-		assert.Equal(renderType, options.Type)
-		assert.Equal("", options.TemplateFile)
-		assert.Equal(file, options.OutputFile)
+		require.Equal(t, renderType, options.Type)
+		require.Equal(t, file, options.OutputFile)
+		require.Empty(t, options.TemplateFile)
 	}
 }
 
-func (assert *PluginTest) TestParseOptionsForCustomTemplate() {
+func TestParseOptionsForCustomTemplate(t *testing.T) {
 	req := new(plugin_go.CodeGeneratorRequest)
 	req.Parameter = proto.String("/path/to/template.tmpl,/base/name/only/output.md")
 
-	options, err := gendoc.ParseOptions(req)
-	assert.Nil(err)
+	options, err := ParseOptions(req)
+	require.NoError(t, err)
 
-	assert.Equal(gendoc.RenderTypeHTML, options.Type)
-	assert.Equal("/path/to/template.tmpl", options.TemplateFile)
-	assert.Equal("output.md", options.OutputFile)
+	require.Equal(t, RenderTypeHTML, options.Type)
+	require.Equal(t, "/path/to/template.tmpl", options.TemplateFile)
+	require.Equal(t, "output.md", options.OutputFile)
 }
 
-func (assert *PluginTest) TestParseOptionsForExcludePatterns() {
+func TestParseOptionsForExcludePatterns(t *testing.T) {
 	req := new(plugin_go.CodeGeneratorRequest)
 	req.Parameter = proto.String(":google/*,notgoogle/*")
 
-	options, err := gendoc.ParseOptions(req)
-	assert.Nil(err)
+	options, err := ParseOptions(req)
+	require.NoError(t, err)
+	require.Len(t, options.ExcludePatterns, 2)
 
-	assert.Equal(2, len(options.ExcludePatterns))
 	pattern0, _ := regexp.Compile("google/*")
 	pattern1, _ := regexp.Compile("notgoogle/*")
-	assert.Equal(pattern0.String(), options.ExcludePatterns[0].String())
-	assert.Equal(pattern1.String(), options.ExcludePatterns[1].String())
+	require.Equal(t, pattern0.String(), options.ExcludePatterns[0].String())
+	require.Equal(t, pattern1.String(), options.ExcludePatterns[1].String())
 }
 
-func (assert *PluginTest) TestParseOptionsWithInvalidValues() {
+func TestParseOptionsWithInvalidValues(t *testing.T) {
 	badValues := []string{
 		"markdown",
 		"html",
@@ -81,42 +72,40 @@ func (assert *PluginTest) TestParseOptionsWithInvalidValues() {
 		req := new(plugin_go.CodeGeneratorRequest)
 		req.Parameter = proto.String(value)
 
-		_, err := gendoc.ParseOptions(req)
-		assert.NotNil(err)
+		_, err := ParseOptions(req)
+		require.Error(t, err)
 	}
 }
 
-func (assert *PluginTest) TestRunPluginForBuiltinTemplate() {
+func TestRunPluginForBuiltinTemplate(t *testing.T) {
 	req := new(plugin_go.CodeGeneratorRequest)
 	req.Parameter = proto.String("markdown,/base/name/only/output.md")
 
-	plugin := new(gendoc.Plugin)
+	plugin := new(Plugin)
 	resp, err := plugin.Generate(req)
-	assert.Nil(err)
-
-	assert.Equal(1, len(resp.File))
-	assert.Equal("output.md", resp.File[0].GetName())
-	assert.NotEmpty(resp.File[0].GetContent())
+	require.NoError(t, err)
+	require.Len(t, resp.File, 1)
+	require.Equal(t, "output.md", resp.File[0].GetName())
+	require.NotEmpty(t, resp.File[0].GetContent())
 }
 
-func (assert *PluginTest) TestRunPluginForCustomTemplate() {
+func TestRunPluginForCustomTemplate(t *testing.T) {
 	req := new(plugin_go.CodeGeneratorRequest)
 	req.Parameter = proto.String("resources/html.tmpl,/base/name/only/output.html")
 
-	plugin := new(gendoc.Plugin)
+	plugin := new(Plugin)
 	resp, err := plugin.Generate(req)
-	assert.Nil(err)
-
-	assert.Equal(1, len(resp.File))
-	assert.Equal("output.html", resp.File[0].GetName())
-	assert.NotEmpty(resp.File[0].GetContent())
+	require.NoError(t, err)
+	require.Len(t, resp.File, 1)
+	require.Equal(t, "output.html", resp.File[0].GetName())
+	require.NotEmpty(t, resp.File[0].GetContent())
 }
 
-func (assert *PluginTest) TestRunPluginWithInvalidOptions() {
+func TestRunPluginWithInvalidOptions(t *testing.T) {
 	req := new(plugin_go.CodeGeneratorRequest)
 	req.Parameter = proto.String("html")
 
-	plugin := new(gendoc.Plugin)
+	plugin := new(Plugin)
 	_, err := plugin.Generate(req)
-	assert.NotNil(err)
+	require.Error(t, err)
 }