+++ CHANGELOG.md
@@ -5,9 +5,17 @@ All noteworthy changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) (as of Feb 2018)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/pseudomuto/protoc-gen-doc/compare/v1.2.0...master)
+## [Unreleased](https://github.com/pseudomuto/protoc-gen-doc/compare/v1.3.0...master)
+
+### Changed
+
+* Switched from dep to go modules
+
+## [v1.3.0](https://github.com/pseudomuto/protoc-gen-doc/compare/v1.2.0...v1.3.0) - April 15, 2019
 
 ### Added
+
+* Support for options and extended options in files, services, methods, enums, enum values, messages and fields [#376](https://github.com/pseudomuto/protoc-gen-doc/pull/376)
 
 ### Changed